// upvertice-ui/src/app/services/models/sponsorship-response.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { ProvidershipLightResponse } from '../models/providership-light-response';
import { SponsorAdResponse } from '../models/sponsor-ad-response';
import { SponsorOfferResponse } from '../models/sponsor-offer-response';
import { SupplierTransactionLightResponse } from '../models/supplier-transaction-light-response';
export interface SponsorshipResponse {
  createdDate?: string;
  id?: number;
  providerships?: Array<ProvidershipLightResponse>;
  sponsorAds?: Array<SponsorAdResponse>;
  sponsorOffer?: SponsorOfferResponse;
  status?: 'PENDING' | 'APPROVED' | 'REJECTED' | 'FINISHED';
  supplierTransactions?: Array<SupplierTransactionLightResponse>;
  userId?: string;
}